requirements.txt
opencv-python~=4.5.1.48
numpy~=1.20.1
pyscreenshot~=2.3
Pillow~=8.1.2
PyChromaKey~=0.0.5